Issue #3872 - move ContainerDefaultConfigurator to config package
