+++ jetty-websocket/javax-websocket-server/src/main/java/module-info.java
@@ -20,9 +20,9 @@ import javax.servlet.ServletContainerInitializer;
 import javax.websocket.server.ServerEndpointConfig;
 
 import org.eclipse.jetty.webapp.Configuration;
+import org.eclipse.jetty.websocket.javax.server.config.ContainerDefaultConfigurator;
 import org.eclipse.jetty.websocket.javax.server.config.JavaxWebSocketConfiguration;
 import org.eclipse.jetty.websocket.javax.server.config.JavaxWebSocketServletContainerInitializer;
-import org.eclipse.jetty.websocket.javax.server.internal.ContainerDefaultConfigurator;
 
 module org.eclipse.jetty.websocket.javax.server
 {
