+++ jetty-websocket/javax-websocket-server/src/main/java/org/eclipse/jetty/websocket/javax/server/config/ContainerDefaultConfigurator.java
@@ -16,7 +16,7 @@
 //  ========================================================================
 //
 
-package org.eclipse.jetty.websocket.javax.server.internal;
+package org.eclipse.jetty.websocket.javax.server.config;
 
 import java.util.List;
 import java.util.ServiceLoader;
