+++ jetty-websocket/javax-websocket-server/src/main/resources/META-INF/services/javax.websocket.server.ServerEndpointConfig$Configurator
@@ -1,1 +1,1 @@
-org.eclipse.jetty.websocket.javax.server.internal.ContainerDefaultConfigurator
+org.eclipse.jetty.websocket.javax.server.config.ContainerDefaultConfigurator
